// backend/src/modules/events/application/event_service.d.ts
export declare const getEventById: (id: number) => Promise<{
    id: number;
    name: string;
    date: Date;
    startHour: Date;
    endHour: Date;
    spaceUsed: string;
    ticketPriceMember: string;
    ticketPriceGuest: string | null;
    capacity: number;
    urlImage: string | null;
    isActive: boolean;
    description: string;
    allowOutsiders: boolean;
    numberOfAssistants: number;
    reservationId: number;
}>;
export declare const createEvent: (data: unknown) => Promise<{
    id: number;
    name: string;
    date: Date;
    startHour: Date;
    endHour: Date;
    spaceUsed: string;
    ticketPriceMember: string;
    ticketPriceGuest: string | null;
    capacity: number;
    urlImage: string | null;
    isActive: boolean;
    description: string;
    allowOutsiders: boolean;
    numberOfAssistants: number;
    reservationId: number;
}>;
export declare const updateEvent: (id: number, data: Partial<CreateEvent>) => Promise<{
    id: number;
    name: string;
    date: Date;
    startHour: Date;
    endHour: Date;
    spaceUsed: string;
    ticketPriceMember: string;
    ticketPriceGuest: string | null;
    capacity: number;
    urlImage: string | null;
    isActive: boolean;
    description: string;
    allowOutsiders: boolean;
    numberOfAssistants: number;
    reservationId: number;
}>;
export declare const getFilteredPaginated: (filter: EventFilter) => Promise<{
    eventos: {
        id: number;
        name: string;
        date: Date;
        startHour: Date;
        endHour: Date;
        spaceUsed: string;
        ticketPriceMember: string;
        ticketPriceGuest: string | null;
        capacity: number;
        urlImage: string | null;
        isActive: boolean;
        description: string;
        allowOutsiders: boolean;
        numberOfAssistants: number;
        reservationId: number;
    }[];
    totalPages: number;
}>;
import { reservationInsertSchema } from "../../../db/schema/Reservation";
export declare const createNewEvent: (dataReservation: typeof reservationInsertSchema._input, dataEvent: typeof eventInsertSchemaM._input) => Promise<{
    reservationInsertId: number;
    eventInsertId: number;
}>;
// ...
export declare const getAllInscriptions: (idEvento: number) => Promise<{
    eventInscriptionId: number;
    nombre: string;
    apellido: string;
    fechaPago: Date;
    dni: string | null;
}[]>;
export declare const deleteEventInscription: (id: number) => Omit<import("drizzle-orm/mysql-core").MySqlUpdateBase<import("drizzle-orm/mysql-core").MySqlTableWithColumns<{
    name: "event_inscription";
    schema: undefined;
    columns: {
        id: import("drizzle-orm/mysql-core").MySqlColumn<{
            name: "id";
            tableName: "event_inscription";
            dataType: "number";
            columnType: "MySqlSerial";
            data: number;
            driverParam: number;
            notNull: true;
            hasDefault: true;
            isPrimaryKey: true;
            isAutoincrement: true;
            hasRuntimeDefault: false;
            enumValues: undefined;
            baseColumn: never;
            identity: undefined;
            generated: undefined;
        }, {}, {}>;
        assisted: import("drizzle-orm/mysql-core").MySqlColumn<{
            name: "assisted";
            tableName: "event_inscription";
            dataType: "boolean";
            columnType: "MySqlBoolean";
            data: boolean;
            driverParam: number | boolean;
            notNull: true;
            hasDefault: false;
            isPrimaryKey: false;
            isAutoincrement: false;
            hasRuntimeDefault: false;
            enumValues: undefined;
            baseColumn: never;
            identity: undefined;
            generated: undefined;
        }, {}, {}>;
        isCancelled: import("drizzle-orm/mysql-core").MySqlColumn<{
            name: "is_cancelled";
            tableName: "event_inscription";
            dataType: "boolean";
            columnType: "MySqlBoolean";
            data: boolean;
            driverParam: number | boolean;
            notNull: true;
            hasDefault: false;
            isPrimaryKey: false;
            isAutoincrement: false;
            hasRuntimeDefault: false;
            enumValues: undefined;
            baseColumn: never;
            identity: undefined;
            generated: undefined;
        }, {}, {}>;
        eventId: import("drizzle-orm/mysql-core").MySqlColumn<{
            name: "events_id";
            tableName: "event_inscription";
            dataType: "number";
            columnType: "MySqlBigInt53";
            data: number;
            driverParam: string | number;
            notNull: true;
            hasDefault: false;
            isPrimaryKey: false;
            isAutoincrement: false;
            hasRuntimeDefault: false;
            enumValues: undefined;
            baseColumn: never;
            identity: undefined;
            generated: undefined;
        }, {}, {}>;
        inscriptionXUser: import("drizzle-orm/mysql-core").MySqlColumn<{
            name: "inscription_x_user";
            tableName: "event_inscription";
            dataType: "number";
            columnType: "MySqlBigInt53";
            data: number;
            driverParam: string | number;
            notNull: true;
            hasDefault: false;
            isPrimaryKey: false;
            isAutoincrement: false;
            hasRuntimeDefault: false;
            enumValues: undefined;
            baseColumn: never;
            identity: undefined;
            generated: undefined;
        }, {}, {}>;
    };
    dialect: "mysql";
}>, import("drizzle-orm/mysql2").MySql2QueryResultHKT, import("drizzle-orm/mysql2").MySql2PreparedQueryHKT, false, "where">, "where">;